Move files under module dir
