+++ my_project_name/config.py
@@ -4,7 +4,7 @@ import os
 import yaml
 import sys
 from typing import List, Any
-from errors import ConfigError
+from my_project_name.errors import ConfigError
 
 logger = logging.getLogger()
